refactor: remove duplicates and clean structure
